refactor(errors): move news_controller.rb error handlers to a service
