+++ backend/app/controllers/api/news_controller.rb
@@ -14,21 +14,14 @@ module Api
   #   "country": "ar"
   # }
   class NewsController < ApplicationController
-    before_action :set_up_client
+    include ErrorHandler
 
-    HEADLINES_API = 'top-headlines'
-    API_URL       = ENV.fetch 'NEWS_API_URL', 'default_url_here'
-    API_KEY       = ENV.fetch 'NEWS_API_KEY', 'default_key_here'
-
-    rescue_from Faraday::Error, with: :handle_faraday_error
-    rescue_from JSON::ParserError, with: :handle_json_error
+    before_action :prepare_index_params, only: :index
 
     # Fetch and return the top headlines for the specified country
     # GET /api/news
     def index
-      country = news_params[:country] || 'ar'
-
-      news = fetch_news country.to_s
+      news = News::Service.call :index, @index_params
 
       render json: { news: }
     end
@@ -36,39 +29,8 @@ module Api
     private
 
     # Strong parameter handling for news
-    def news_params
-      params.permit :country
-    end
-
-    # Fetches the latest headlines for a given country
-    # @param [String] country The country code
-    # @return [Array<Hash>] An array of news articles
-    def fetch_news(country)
-      response = @connection.get do |request|
-        request.url HEADLINES_API
-        request.params['country'] = country
-      end
-
-      raise Faraday::Error, 'Bad response from News API' unless response.success?
-
-      JSON.parse response.body
-    end
-
-    # Sets up the Faraday HTTP client with the necessary configurations
-    def set_up_client
-      @connection = Faraday.new(API_URL) do |faraday|
-        faraday.headers['X-Api-Key'] = API_KEY
-      end
-    end
-
-    # Handles Faraday errors and responds with a 502 Bad Gateway status
-    def handle_faraday_error(exception)
-      render json: { error: "Failed to fetch news: #{exception.message}" }, status: :bad_gateway
-    end
-
-    # Handles JSON parsing errors and responds with a 502 Bad Gateway status
-    def handle_json_error(_exception)
-      render json: { error: 'Failed to parse news data' }, status: :bad_gateway
+    def prepare_index_params
+      @index_params = params.permit :country
     end
   end
 end
